+++ aiogremlin/__init__.py
@@ -1,5 +1,5 @@
 from aiogremlin.driver.cluster import Cluster
 from aiogremlin.remote.driver_remote_connection import DriverRemoteConnection
-from aiogremlin.gremlin_python.structure.graph import Graph
+from aiogremlin.structure.graph import Graph
 
 __version__ = "3.2.4"
