+++ aiogremlin/remote/driver_remote_connection.py
@@ -2,9 +2,9 @@ import asyncio
 from urllib.parse import urlparse
 
 from aiogremlin.driver.cluster import Cluster
-from aiogremlin.gremlin_python.driver import serializer
+from gremlin_python.driver import serializer
 from aiogremlin.remote.driver_remote_side_effects import RemoteTraversalSideEffects
-from aiogremlin.gremlin_python.driver.remote_connection import RemoteTraversal
+from aiogremlin.remote.remote_connection import RemoteTraversal
 
 
 __author__ = 'David M. Brown (davebshow@gmail.com)'
