--- docs/aiogremlin.gremlin_python.driver.rst
@@ -1,45 +1,0 @@
-aiogremlin.gremlin_python.driver package
-========================================
-
-Submodules
-----------
-
-aiogremlin.gremlin_python.driver.protocol module
-------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.driver.protocol
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.driver.remote_connection module
----------------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.driver.remote_connection
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.driver.request module
------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.driver.request
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.driver.serializer module
---------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.driver.serializer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.driver.transport module
--------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.driver.transport
-    :members:
-    :undoc-members:
-    :show-inheritance:
