--- docs/aiogremlin.gremlin_python.process.rst
@@ -1,29 +1,0 @@
-aiogremlin.gremlin_python.process package
-=========================================
-
-Submodules
-----------
-
-aiogremlin.gremlin_python.process.graph_traversal module
---------------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.process.graph_traversal
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.process.strategies module
----------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.process.strategies
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-aiogremlin.gremlin_python.process.traversal module
---------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.process.traversal
-    :members:
-    :undoc-members:
-    :show-inheritance:
