--- docs/aiogremlin.gremlin_python.rst
@@ -1,22 +1,0 @@
-aiogremlin.gremlin_python package
-=================================
-
-Subpackages
------------
-
-.. toctree::
-
-    aiogremlin.gremlin_python.driver
-    aiogremlin.gremlin_python.process
-    aiogremlin.gremlin_python.structure
-
-Submodules
-----------
-
-aiogremlin.gremlin_python.statics module
-----------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.statics
-    :members:
-    :undoc-members:
-    :show-inheritance:
