--- docs/aiogremlin.gremlin_python.structure.io.rst
@@ -1,13 +1,0 @@
-aiogremlin.gremlin_python.structure.io package
-==============================================
-
-Submodules
-----------
-
-aiogremlin.gremlin_python.structure.io.graphson module
-------------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.structure.io.graphson
-    :members:
-    :undoc-members:
-    :show-inheritance:
