--- docs/aiogremlin.gremlin_python.structure.rst
@@ -1,20 +1,0 @@
-aiogremlin.gremlin_python.structure package
-===========================================
-
-Subpackages
------------
-
-.. toctree::
-
-    aiogremlin.gremlin_python.structure.io
-
-Submodules
-----------
-
-aiogremlin.gremlin_python.structure.graph module
-------------------------------------------------
-
-.. automodule:: aiogremlin.gremlin_python.structure.graph
-    :members:
-    :undoc-members:
-    :show-inheritance:
