+++ setup.py
@@ -1,24 +1,37 @@
+import pip
 from setuptools import setup
+from distutils.command.build_py import build_py as _build_py
+
+
+class build_py(_build_py):
+    """Don't install tornado when installing gremlinpython"""
+    pip.main([
+    'install',
+    'git+https://git-wip-us.apache.org/repos/asf/tinkerpop.git@12aa2d2fe534fbd9540c9328725ab844e4bb1010#egg=gremlinpython&subdirectory=gremlin-python/src/main/jython',
+    '--no-deps'])
 
 
 setup(
     name='aiogremlin',
-    version='3.2.4',
+    version='3.2.5rc1',
     url='',
     license='Apache Software License',
     author='davebshow',
     author_email='davebshow@gmail.com',
     description='Async Gremlin-Python',
     long_description=open('README.txt').read(),
-    packages=['aiogremlin', 'aiogremlin.driver', 'aiogremlin.driver.aiohttp',
-              'aiogremlin.gremlin_python', 'aiogremlin.gremlin_python.driver',
-              'aiogremlin.gremlin_python.process',
-              'aiogremlin.gremlin_python.structure',
-              'aiogremlin.gremlin_python.structure.io',
+    packages=['aiogremlin',
+              'aiogremlin.driver',
+              'aiogremlin.driver.aiohttp',
+              'aiogremlin.process',
+              'aiogremlin..structure',
               'aiogremlin.remote'],
+    cmdclass={'build_py': build_py},
     install_requires=[
+        'aenum==1.4.5',  # required gremlinpython dep
         'aiohttp==1.3.3',
-        'PyYAML==3.12'
+        'PyYAML==3.12',
+        'six==1.10.0'  # required gremlinpython dep
     ],
     test_suite='tests',
     setup_requires=['pytest-runner'],
