+++ tests/config/config.yml
@@ -1,4 +1,4 @@
 scheme: 'wss'
 hosts: ['localhost']
 port: 8183
-message_serializer: 'aiogremlin.gremlin_python.driver.serializer.GraphSONMessageSerializer'
+message_serializer: 'gremlin_python.driver.serializer.GraphSONMessageSerializer'
