+++ tests/config_module.py
@@ -1,4 +1,4 @@
 SCHEME = 'wss'
 HOSTS = ['localhost']
 PORT = 8183
-MESSAGE_SERIALIZER = 'aiogremlin.gremlin_python.driver.serializer.GraphSONMessageSerializer'
+MESSAGE_SERIALIZER = 'gremlin_python.driver.serializer.GraphSONMessageSerializer'
