+++ tests/test_connection_protocol.py
@@ -20,7 +20,7 @@
 # import pytest
 #
 # from aiogremlin import exception
-# from aiogremlin.gremlin_python.driver import request, serializer
+# from gremlin_python.driver import request, serializer
 #
 #
 # @pytest.mark.asyncio
