+++ tests/test_gremlin_python/process/test_strategies.py
@@ -22,9 +22,9 @@ __author__ = 'Marko A. Rodriguez (http://markorodriguez.com)'
 import unittest
 from unittest import TestCase
 
-from aiogremlin.gremlin_python.structure.graph import Graph
-from aiogremlin.gremlin_python.process.strategies import *
-from aiogremlin.gremlin_python.process.graph_traversal import __
+from aiogremlin.structure.graph import Graph
+from gremlin_python.process.strategies import *
+from gremlin_python.process.graph_traversal import __
 
 
 class TestTraversalStrategies(TestCase):
